+++ response.go
@@ -198,9 +198,11 @@ func (r *Response) Download(file string, fileName string) error {
 // If debugging is not enabled, only the stauts code is set, which means you can still
 // write to the response, or use your error status handler.
 func (r *Response) Error(err interface{}) error {
-	if r.err == nil {
-		ErrLogger.Println(err)
-	}
+	ErrLogger.Println(err)
+	return r.error(err)
+}
+
+func (r *Response) error(err interface{}) error {
 	r.err = err
 	if config.GetBool("debug") {
 		stacktrace := r.stacktrace